+++ R/chronique.DCE.R
@@ -88,7 +88,7 @@ chronique.DCE <- function(
     mutate(tpcomm_commune_libelle = aquatools::BV.ComByCoordL93(chsta_coord_x,chsta_coord_y) %>% dplyr::select(name) %>% as.character()) %>% 
     ungroup() %>% 
     mutate(chsta_detailsloc = glue('{chsta_rive} - {chsta_detailsloc}')) %>% 
-    dplyr::select(chsta_codesie, chsta_codepointprlvmt, chsta_codemo, chsta_milieu, tpcomm_commune_libelle, chsta_coderhj, chsta_coord_x, chsta_coord_y, chsta_detailsloc, chsta_ombrage, chsta_facies) %>% 
+    dplyr::select(chsta_codesie, chsta_codepointprlvmt, chsta_codemo, chsta_mo, chsta_milieu, tpcomm_commune_libelle, chsta_coderhj, chsta_coord_x, chsta_coord_y, chsta_detailsloc, chsta_ombrage, chsta_facies) %>% 
     {if(nrow(SuiviTerrain) != 0) right_join(., SuiviTerrain %>% 
                  mutate(chsvi_date = format(chsvi_date, format="%d/%m/%Y")) %>% 
                  mutate(chsvi_profondeur = ifelse(chsvi_profondeur > 10, chsvi_profondeur / 10, chsvi_profondeur)) %>%
@@ -107,10 +107,11 @@ chronique.DCE <- function(
                                             by = c('chsvi_capteur' = 'chcap_numerocapteur')
     ) else .} %>% 
     {if(nrow(Capteurs) == 0 | nrow(SuiviTerrain) == 0) mutate(., chcap_modelecapteur = NA_character_) else .} %>% 
-    dplyr::select(chsta_codesie, chsta_codepointprlvmt, chsta_codemo, chsta_milieu, tpcomm_commune_libelle, chsta_coderhj, chsta_coord_x, chsta_coord_y, chsvi_capteur, chcap_modelecapteur, chsvi_date, chsta_detailsloc, chsta_ombrage, chsta_facies, chsvi_profondeur, chsvi_valeur, chsvi_operateurs, chsvi_remarques) %>% 
+    dplyr::select(chsta_codesie, chsta_codepointprlvmt, chsta_codemo, chsta_mo, chsta_milieu, tpcomm_commune_libelle, chsta_coderhj, chsta_coord_x, chsta_coord_y, chsvi_capteur, chcap_modelecapteur, chsvi_date, chsta_detailsloc, chsta_ombrage, chsta_facies, chsvi_profondeur, chsvi_valeur, chsvi_operateurs, chsvi_remarques) %>% 
     rename(CODE_STA_SANDRE = chsta_codesie, 
            CODE_PT_PRELEVT_SANDRE = chsta_codepointprlvmt,
            COD_STA_METIER = chsta_codemo, 
+           OPERATEUR = chsta_mo, 
            RIV_NOM = chsta_milieu, 
            COM_NOM = tpcomm_commune_libelle, 
            NOM_STA_METIER = chsta_coderhj,
@@ -126,6 +127,7 @@ chronique.DCE <- function(
            T_POSE = chsvi_valeur,
            AGENT_POSE = chsvi_operateurs,
            RQ_POSE = chsvi_remarques) %>%
+    filter(!(is.na(COD_STA_METIER) & is.na(RIV_NOM) & is.na(X_93))) %>% # Rustine temporaire pour effacer le suivi qui ne correspond pas à la station : code à reprendre en profondeur
     st_drop_geometry()
 
   ### Chronique ###
@@ -138,7 +140,8 @@ chronique.DCE <- function(
     mutate(chmes_capteur = ifelse("chmes_capteur" %in% names(.), chmes_capteur, NA_character_)) %>% 
     mutate(chmes_validation = ifelse("chmes_validation" %in% names(.), chmes_validation, NA_character_)) %>% 
     mutate(chmes_mode_acquisition = ifelse("chmes_mode_acquisition" %in% names(.), chmes_mode_acquisition, "Mesuré")) %>% 
-    dplyr::select(chsta_codesie, chsta_codepointprlvmt, chsta_codemo, chmes_coderhj, chmes_capteur, chmes_date, chmes_heure, chmes_valeur, chmes_validation, chmes_mode_acquisition) %>% 
+    mutate(chmes_referentiel_temporel = ifelse("chmes_referentiel_temporel" %in% names(.), chmes_referentiel_temporel, NA_character_)) %>% 
+    dplyr::select(chsta_codesie, chsta_codepointprlvmt, chsta_codemo, chmes_coderhj, chmes_capteur, chmes_date, chmes_heure, chmes_referentiel_temporel, chmes_valeur, chmes_validation, chmes_mode_acquisition) %>% 
     rename(CODE_STA_SANDRE = chsta_codesie, 
            CODE_PT_PRELEVT_SANDRE = chsta_codepointprlvmt,
            COD_STA_METIER = chsta_codemo, 
@@ -146,6 +149,7 @@ chronique.DCE <- function(
            COD_SONDE = chmes_capteur, 
            DATE = chmes_date, 
            HEURE = chmes_heure, 
+           REF_HORAIRE = chmes_referentiel_temporel, 
            T = chmes_valeur)
   
   ##### Export à proprement parler #####